Resolve #150 Pick Displayed Stories
This commit refactors the story creation feature to include stories in the administrative panel. This will enable administrators to view and edit stories. It also adds a display attribute to enable users to filter whether a story is displayed in the index by toggling the display to true or not.