Resolve #304 Flexbox Height Calculation
The height of our flexbox container was not correctly calculating in the case where we had four boxes and one was a half height box. This commit refactors the weigh in grid javascript to calculate the height of the grid in pixels and round it up to the nearest box height to avoid 4 column layouts.