Fix Broken Stories Header
The stories header disappeared because its styling was disconnected from its class in 431c2b07ce59175b026460b13bce3f800d9bdf57. This commit fixes this by updating the class used in the page template to the v2-header class that now owns the styling we used on the page.